AB#26873 - Zones - Refactor from ApplicationUiSettingsAppService to ZoneManagementAppService
