--- applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/SettingManagement/IApplicationUiSettingsAppService.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Unity.GrantManager.Zones;
-using Volo.Abp.Application.Services;
-
-namespace Unity.GrantManager.SettingManagement;
-public interface IApplicationUiSettingsAppService : IApplicationService
-{
-    Task<ApplicationUiSettingsDto> GetAsync();
-    Task<ZoneGroupDefinitionDto?> GetForFormAsync(Guid formId);
-    Task UpdateAsync(string providerName, string providerKey, List<UpdateZoneDto> input);
-    Task SetConfigurationAsync();
-    Task SetConfigurationAsync(Guid formId, ZoneGroupDefinitionDto? input);
-}
