+++ applications/Unity.GrantManager/src/Unity.GrantManager.Domain/Zones/IZoneManager.cs
@@ -6,7 +6,8 @@ namespace Unity.GrantManager.Zones;
 public interface IZoneManager
 {
     Task<ZoneGroupDefinition> GetAsync();
-    Task<ZoneGroupDefinition> GetAsync(Guid formId);
+    Task<ZoneGroupDefinition> GetAsync(string providerKey);
+    Task<ZoneGroupDefinition> GetAsync(string providerName, string providerKey);
     Task SetForFormAsync(Guid formId, ZoneGroupDefinition template);
     Task SetForTennantAsync(ZoneGroupDefinition template);
 }
