+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Components/ApplicationUiSettingGroup/ApplicationUiSettingGroupViewComponent.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 using Unity.GrantManager.SettingManagement;
+using Unity.GrantManager.Zones;
 using Volo.Abp.AspNetCore.Mvc;
 
 namespace Unity.GrantManager.Web.Components.ApplicationUiSettingGroup;
 
-public class ApplicationUiSettingGroupViewComponent(IApplicationUiSettingsAppService settingsAppService) : AbpViewComponent
+public class ApplicationUiSettingGroupViewComponent(IZoneManagementAppService settingsAppService) : AbpViewComponent
 {
     public virtual async Task<IViewComponentResult> InvokeAsync()
     {
